+++ src/script/sign.h
@@ -9,6 +9,7 @@
 #include <hash.h>
 #include <pubkey.h>
 #include <script/interpreter.h>
+#include <script/keyorigin.h>
 #include <streams.h>
 
 class CKey;
@@ -18,28 +19,6 @@ class CScriptID;
 class CTransaction;
 
 struct CMutableTransaction;
-
-struct KeyOriginInfo
-{
-    unsigned char fingerprint[4]; //!< First 32 bits of the Hash160 of the public key at the root of the path
-    std::vector<uint32_t> path;
-
-    friend bool operator==(const KeyOriginInfo& a, const KeyOriginInfo& b)
-    {
-        return std::equal(std::begin(a.fingerprint), std::end(a.fingerprint), std::begin(b.fingerprint)) && a.path == b.path;
-    }
-
-    SERIALIZE_METHODS(KeyOriginInfo, obj)
-    {
-        READWRITE(obj.fingerprint, obj.path);
-    }
-
-    void clear()
-    {
-        memset(fingerprint, 0, 4);
-        path.clear();
-    }
-};
 
 /** An interface to be implemented by keystores that support signing. */
 class SigningProvider
